feat(MapLayout): Add MapLayout
